misc/HtmlTestSuite/results/test-0914.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>document.write</title>
<script src="/resources/testharness.js"/><script src="/resources/testharnessreport.js"/>
<script>
var t = async_test();
t.step(function() {
  var s = "&lt;script src='013.js'&gt;&lt;" + "/script&gt;&lt;/svg&gt;]]&gt;&lt;path&gt;&lt;/svg&gt;";
  for (var i=0; i&lt;s.length; i++) {
    document.write(s[i]);
  }
});
</script><script>
t.step(function() {
  assert_equals(document.body.childNodes[0].nodeType, document.ELEMENT_NODE);
  assert_equals(document.body.childNodes[0].localName, "svg");
  assert_equals(document.body.childNodes[0].childNodes[0].nodeType, document.TEXT_NODE);
  assert_equals(document.body.childNodes[0].childNodes[0].data, "&lt;/svg&gt;");
  assert_equals(document.body.childNodes[0].childNodes[1].nodeType, document.ELEMENT_NODE);
  assert_equals(document.body.childNodes[0].childNodes[1].localName, "path");
}
);
t.done();
</script>
</head><body><div id="log"/>
</body></html>
